+++ src/Bullet3Serialize/Bullet2FileLoader/b3Serializer.h
@@ -305,16 +305,16 @@ protected:
 
 			if (!littleEndian)
 			{
-				shtPtr[0] = (int)b3SwapEndian(shtPtr[0]);
-				shtPtr[1] = (int)b3SwapEndian(shtPtr[1]);
+				shtPtr[0] = (short)b3SwapEndian(shtPtr[0]);
+				shtPtr[1] = (short)b3SwapEndian(shtPtr[1]);
 
 				int len = shtPtr[1];
 				shtPtr += 2;
 
 				for (int a = 0; a < len; a++, shtPtr += 2)
 				{
-					shtPtr[0] = (int)b3SwapEndian(shtPtr[0]);
-					shtPtr[1] = (int)b3SwapEndian(shtPtr[1]);
+					shtPtr[0] = (short)b3SwapEndian(shtPtr[0]);
+					shtPtr[1] = (short)b3SwapEndian(shtPtr[1]);
 				}
 			}
 			else
